Moving terminal size to seperate file, preparing for progress bar addition
